Fix for Restart enable/disable in Debug menu dropdown
The behavior for enabling and disabling the restart selection in the Debug menu dropdown was different than the restart button in the toolbar.  The toolbar uses the mSimulationStopped variable.  The dropdown menu used the simulatorReady variable.  The toolbar seemed to exhibit the correct behavior, so the dropdown menu was changed to match the toolbar.